大臣对话历史改回 agno session + 跨月 LLM 纪要
月内对话历史交回 agno 每月一个 session 自管（含 tool 调用痕迹），
解决大臣空转重复拟旨（如王承恩反复拟孙承宗入阁）：自管 chat_messages
只存纯文本、丢了 tool 痕迹，大臣看不到自己拟的旨真入了档。

- session.py/web_app.py：删自管历史前置旁路，不再传 add_history_to_context=False；
  流式路径补 draft_line 注入；chat_messages 保留只做前端展示。
- 跨月失忆由月末 LLM 压缩的「大臣私人对话纪要」补：新 prompt minister_recap.md
  + create_minister_recap_agent + record_minister_recaps（懒生成，只跑本月召见过的大臣）
  + DB 存取（复用 event_memories，subject_id=大臣名，按大臣私有）
  + build_minister_recap_brief 月初注入大臣 system。

同时并入此前未提交的设定/db/前端等改动。

+++ ming_sim/db/admin.py
@@ -89,7 +89,7 @@ class _AdminMixin:
         if table == "metrics" and data.get("key") in ("国库", "内库") and "value" in data:
             self.conn.execute(
                 "UPDATE economy_accounts SET balance = ? WHERE account = ?",
-                (int(data["value"]), data["key"]),
+                (float(data["value"]), data["key"]),
             )
         self.conn.commit()
         row = self.conn.execute(f"SELECT * FROM {table} WHERE {pk}=?", (data[pk],)).fetchone()
